+++ keyword_reference_test.go
@@ -16,6 +16,7 @@ func TestRecordedKeywordPrograms(t *testing.T) {
 		"keyword-faca-accented", "keyword-funcao-accented",
 		"alias-caracter-lookalike",
 		"division-alias-values", "division-alias-case", "division-alias-order",
+		"function-ending-accented", "function-ending-uppercase",
 	} {
 		t.Run(id, func(t *testing.T) {
 			dir := filepath.Join("testdata/conformance/visualg-3.0.7/probes", id)